refactor(f): update page logic
