+++ packages/frontend/src/routes/(app)/category/[category=categories]/[slug]/+layout.server.ts
@@ -1,102 +1,28 @@
-import { error, type ServerLoadEvent } from '@sveltejs/kit';
-import type { MetaTagsProps } from 'svelte-meta-tags';
-import { site } from '$lib/variables';
-import { createAuthorLink, createAuthorString } from '$lib/helpers';
 import type { Article } from '$lib/schema';
-import type { LayoutServerLoad } from './$types';
+import type { LayoutServerLoad, LayoutServerLoadEvent } from './$types';
 
-export const load: LayoutServerLoad = (async (event: ServerLoadEvent) => {
+export const load: LayoutServerLoad = (async (event: LayoutServerLoadEvent) => {
 	const { category, slug } = event.params;
 
 	/**
-	 * Retrieve article information.
+	 * Query for the article here rather than in `+page.server.ts` because
+	 * we want to have article data also included in `+layout.svelte`. This
+	 * method ensures that whatever is rendered in `+page.svelte` is truly
+	 * only the article and nothing else.
 	 */
-
 	const req = await event.fetch(`/api/article?category=${category}&slug=${slug}`);
 	const article: Article | undefined = await req.json();
 
-	/**
-	 * Build article information.
-	 */
-
-	if (article) {
-		const title = article.title;
-		const subtitle = article.subtitle
-			? article.subtitle
-			: `An article by ${createAuthorString(article.authors)} at ${site.title}`;
-
-		let ogTitle = title;
-		let ogDescription = subtitle;
-		let ogTags = [
-			// Add default site tags.
-			...site.tags,
-
-			// Add the article's tags.
-			...(article.tags
-				? article.tags.map((t) => {
-						return t.name;
-					})
-				: [])
-		];
-
-		// If metaInfo is not null, that means there is data
-		// we can customize in the HTML document.
-		if (article.metaInfo) {
-			if (article.metaInfo.ogTitle) {
-				ogTitle = article.metaInfo.ogTitle;
-			}
-
-			if (article.metaInfo.ogTags) {
-				ogTags.push(...article.metaInfo.ogTags);
-			}
-
-			if (article.metaInfo.ogDescription) {
-				ogDescription = article.metaInfo.ogDescription;
-			}
-
-			if (article.metaInfo.ogImage) {
-				// TODO
-			}
-		}
-
-		// Create an array of links to author's profile pages.
-		const ogAuthorLinks = [
-			...article.authors.map((n) => {
-				return createAuthorLink(site.url, n.slug.current);
-			})
-		];
-
-		const pageMetaTags = Object.freeze({
-			description: ogDescription,
-			openGraph: {
-				title: ogTitle,
-				description: ogDescription,
-				type: 'article',
-				article: {
-					// Article dates in ISO-8601 format
-					publishedTime: `${article.date}T00:00:00Z`,
-					modifiedTime: `${article.updatedDate ? article.updatedDate : article.date}T00:00:00Z`,
-					authors: [...ogAuthorLinks],
-					tags: ogTags,
-					section: article.category.name
-				}
-			},
-			twitter: {
-				title: ogTitle,
-				description: ogDescription
-			}
-		}) satisfies MetaTagsProps;
+	return {
+		article,
 
 		/**
-		 * Return page data.
+		 * This is used for loading related article data for page slugs. Also,
+		 * make sure that this field is declared AFTER article. This ensures
+		 * that the more pertinent data, which is the article data, is requested
+		 * first and foremost, rather than non-critical data like related
+		 * articles.
 		 */
-
-		return {
-			article,
-			title,
-			pageMetaTags
-		};
-	}
-
-	error(404, 'Article not found...');
+		parentData: await event.parent()
+	};
 }) satisfies LayoutServerLoad;
