+++ packages/frontend/src/routes/(app)/category/[category=categories]/[slug]/+layout.svelte
@@ -1,11 +1,6 @@
 <script lang="ts">
 	import type { Snippet } from 'svelte';
 	import type { LayoutData } from './$types';
-	import Subtitle from '$components/defaults/Subtitle.svelte';
-	import PhotoCaption from '$components/custom/PhotoCaption.svelte';
-	import Image from '$components/Image.svelte';
-	import ByLine from '$components/article/ByLine.svelte';
-	import DateLine from '$components/article/DateLine.svelte';
 
 	interface Props {
 		data: LayoutData;
@@ -13,90 +8,36 @@
 	}
 
 	let { data, children }: Props = $props();
-	let title = $derived(data.article.title);
-	let subtitle = $derived(data.article.subtitle);
-	let media = $derived(data.article.media);
-	let blurHash = $derived(data.article.asset?.metadata.blurHash);
-	let authors = $derived(data.article.authors);
-	let date = $derived(data.article.date);
-	let series = $derived(data.article.series);
-	let category = $derived(data.article.category);
-	let updatedDate = $derived(data.article.updatedDate);
+	let relatedArticles = $derived(data.parentData.articles);
+	let articleCategory = $derived(data.article?.category.name);
+	let articleTitle = $derived(data.article?.title);
 </script>
 
+<ol aria-label="breadcrumbs" class="breadcrumb text-sm">
+	<li>{articleCategory}</li>
+	<li aria-current="page">{articleTitle}</li>
+</ol>
+
 <article class="m-0 p-0 md:m-1 md:p-1 lg:m-2 lg:p-2 relative w-full sm:max-w-5xl center">
-	<header class="flex flex-col center">
-		<div class="pb-1 mb-1 lg:mb-1 lg:pb-1 w-fit">
-			{#await title}
-				Loading title...
-			{:then title}
-				{#if title}
-					<h1
-						class="text-4xl sm:text-5xl lg:text-7xl font-display font-black font-stretch-condensed tracking-tight ml-1 p-2 sm:mb-3 sm:pb-3 sm:leading-14 lg:leading-18 antialiased"
-					>
-						{title}
-					</h1>
-				{/if}
-			{/await}
-			{#if subtitle}
-				<div class="mb-8 max-w-3xl">
-					<Subtitle>
-						{subtitle}
-					</Subtitle>
-				</div>
-			{/if}
-		</div>
-		{#if media}
-			<div class="w-full center">
-				<figure role="group" class="mb-1 pb-1 sm:pb-4 sm:mb-4 center">
-					<div class="mb-2">
-						<Image
-							{media}
-							alt={media.alt}
-							width={1920}
-							height={1080}
-							priority={true}
-							{blurHash}
-							loading={'eager'}
-						/>
-					</div>
-					{#if data.article.asset!.creditLine}
-						<figcaption class="sm:p-1 pt-1 p-3">
-							<PhotoCaption>
-								{data.article.asset!.creditLine}
-							</PhotoCaption>
-						</figcaption>
-					{/if}
-				</figure>
-			</div>
-		{/if}
-		<div class="flex flex-col mb-4 ml-1 pl-2 sm:ml-4 sm:pl-4">
-			<div class="flex flex-row items-baseline align-center">
-				<ByLine {authors} />&nbsp;
-				<span class="font-bold text-sm">✍&nbsp;</span>
-				{#if series}
-					<a
-						class="italic font-serif text-sm font-bold"
-						href="/series/{series.slug.current}"
-						title="{series.name} series"
-					>
-						{series.name}
-					</a>
-				{:else}
-					<a
-						class="text-sm tracking-wider font-semibold hover:underline"
-						href="/category/{category.slug.current}"
-						title={category.name}
-					>
-						{category.name}
-					</a>
-				{/if}
-			</div>
-			<DateLine {date} locale={data.userLocale} />
-			{#if data.article.updatedDate}
-				<DateLine date={updatedDate} locale={data.userLocale} updated={true} />
-			{/if}
-		</div>
-	</header>
 	{@render children()}
 </article>
+
+<aside>
+	<h1>Related Articles</h1>
+	{relatedArticles}
+</aside>
+
+<style>
+	.breadcrumb {
+		display: flex;
+		flex-wrap: wrap;
+		list-style: none;
+		margin: 0;
+		padding: 0;
+		align-items: end;
+	}
+
+	.breadcrumb li:not(:first-child)::before {
+		content: '→';
+	}
+</style>
